move method from authenService to otpService
