+++ README.md
@@ -16,10 +16,11 @@
 - Redis
 - Docker
 - Lombok, MapStruct, Mapper
+- 
 
 # 🧱 Quy ước tạo file mới trong dự án: tạo các thư để chứa những file không phải global
   # Tuân thủ các quy tắc đặt tên:
-    - class: Invoice aaaaaa
+    - class: Invoice
     - folder: invoice 
     - loại: 
      + biến đơn: camelCase, rõ chức năng --VD: fullName
