+++ src/main/java/com/example/datn_qlnt_manager/common/Gender.java
@@ -3,5 +3,5 @@ package com.example.datn_qlnt_manager.common;
 public enum Gender {
     MALE,
     FEMALE,
-    OTHER,
+    UNKNOWN,
 }
